+++ src/dns_exporter/metrics.py
@@ -51,6 +51,7 @@ FAILURE_REASONS = [
     "invalid_request_query_name",
     "invalid_request_query_type",
     "invalid_request_query_class",
+    "certificate_error",
     "connection_error",
     "connection_refused",
     "timeout",